Modularized app away from react render
Index.jsx was performing multiple duties. This refactor separates out those responsibilities by splitting the render into the html (index.jsx) from the component details (app.jsx).
